requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.0.4
astro-sdk-python==1.6.1
duckdb
pendulum
numpy
matplotlib
jupyter
ipykernel
ipywidgets
plotly
